+++ srcjs/input_binding_slider.js
@@ -130,6 +130,14 @@ $.extend(sliderInputBinding, textInputBinding, {
       }
     }
 
+    // Recalculate the number of ticks
+    var max = typeof msg.max === 'undefined' ? slider.options.max : +msg.max;
+    var min =  typeof msg.min === 'undefined' ? slider.options.min : +msg.min;
+    var step = typeof msg.step === 'undefined' ? slider.options.step : +msg.step;
+    var nSteps = (max - min) / step;
+    var scaleFactor = Math.ceil(nSteps / 10);
+    msg.grid_num = nSteps / scaleFactor;
+
     updateLabel(data.label, this._getLabelNode(el));
 
     var domElements = ['data-type', 'time-format', 'timezone'];