+++ NEWS.md
@@ -41,6 +41,8 @@ shiny 1.3.2.9001
 
 * Fixed [#2329](https://github.com/rstudio/shiny/issues/2329), [#1817](https://github.com/rstudio/shiny/issues/1817): These bugs were reported as fixed in Shiny 1.3.0 but were not actually fixed because some JavaScript changes were accidentally not included in the release. The fix resolves issues that occur when `withProgressBar()` or bookmarking are combined with the [networkD3](https://christophergandrud.github.io/networkD3/) package's Sankey plot.
 
+* Fixed [#2398](https://github.com/rstudio/shiny/issues/2398): `updateSliderInput` failed to update the number of ticks.
+
 shiny 1.3.2
 ===========